+++ R/utils_link_inverse.R
@@ -11,3 +11,19 @@ get_link_fun <- function(model) {
 
   NULL
 }
+
+
+#' @importFrom stats terms weights
+get_model_weights <- function(model) {
+  w <- NULL
+  tryCatch(
+    {
+      w <- stats::weights(model)
+    },
+    error = function(x) { NULL },
+    warning = function(x) { NULL },
+    finally = function(x) { NULL }
+  )
+
+  w
+}